requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.2
django-environ==0.13.0
django-import-export==4.4.0
djangorestframework==3.16.1
sqlparse==0.5.5
tablib==3.9.0
